LPD-54772 Move structure builder services to the top folder
