+++ modules/apps/site/site-cms-site-initializer/src/main/resources/META-INF/resources/js/main/categorization/services/SpaceService.ts
@@ -3,7 +3,7 @@
  * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
  */
 
-import ApiHelper from '../../../structure_builder/services/ApiHelper';
+import ApiHelper from '../../../services/ApiHelper';
 
 async function getSpaces(): Promise<
 	{
