+++ modules/apps/site/site-cms-site-initializer/src/main/resources/META-INF/resources/js/structure_builder/components/StructureBuilderManagementBar.tsx
@@ -11,6 +11,7 @@ import {ManagementToolbar, openToast} from 'frontend-js-components-web';
 import {navigate} from 'frontend-js-web';
 import React, {Dispatch} from 'react';
 
+import StructureService from '../../services/StructureService';
 import {config} from '../config';
 import {
 	Action,
@@ -29,7 +30,6 @@ import selectStructureName from '../selectors/selectStructureName';
 import selectStructureSpaces from '../selectors/selectStructureSpaces';
 import selectStructureStatus from '../selectors/selectStructureStatus';
 import selectUnsavedChanges from '../selectors/selectUnsavedChanges';
-import StructureService from '../services/StructureService';
 import {useValidate} from '../utils/validation';
 import AsyncButton from './AsyncButton';
 import ManagementBar from './ManagementBar';
