+++ modules/apps/site/site-cms-site-initializer/src/main/resources/META-INF/resources/js/structure_builder/components/picklist_builder/PicklistBuilderManagementBar.tsx
@@ -16,7 +16,7 @@ import {
 	useOptions,
 	useSetId,
 } from '../../contexts/PicklistBuilderContext';
-import PicklistService from '../../services/PicklistService';
+import PicklistService from '../../../services/PicklistService';
 import focusInvalidElement from '../../utils/focusInvalidElement';
 import AsyncButton from '../AsyncButton';
 import ManagementBar from '../ManagementBar';
